+++ test/lib/Elastica/Test/Exception/Connection/MemcacheExceptionTest.php
@@ -1,8 +1,0 @@
-<?php
-namespace Elastica\Test\Exception\Connection;
-
-use Elastica\Test\Exception\AbstractExceptionTest;
-
-class MemcacheExceptionTest extends AbstractExceptionTest
-{
-}